+++ lib/src/http/serializable.dart
@@ -7,7 +7,7 @@ import 'http.dart';
 
 /// Interface for serializable instances to be decoded from an HTTP request body and encoded to an HTTP response body.
 ///
-/// Implementers of this interface may be a [Response.body] and bound with an [Bind.body] in [RESTController].
+/// Implementers of this interface may be a [Response.body] and bound with an [Bind.body] in [ResourceController].
 abstract class HTTPSerializable {
   /// Documents [serializable].
   ///
@@ -40,7 +40,7 @@ abstract class HTTPSerializable {
 
   /// Reads values from [requestBody] into an object.
   ///
-  /// This method is invoked when an [RESTController] property or operation method argument is bound with [Bind.body]. [requestBody] is the
+  /// This method is invoked when an [ResourceController] property or operation method argument is bound with [Bind.body]. [requestBody] is the
   /// request body of the incoming HTTP request, decoded according to its content-type.
   void readFromMap(Map<String, dynamic> requestBody);