+++ test/http/resource_controller_body_binding_test.dart
@@ -196,28 +196,28 @@ class CrashModel extends HTTPSerializable {
   }
 }
 
-class TestController extends RESTController {
+class TestController extends ResourceController {
   @Operation.post()
   Future<Response> create(@Bind.body() TestModel tm) async {
     return new Response.ok(tm);
   }
 }
 
-class ListTestController extends RESTController {
+class ListTestController extends ResourceController {
   @Operation.post()
   Future<Response> create(@Bind.body() List<TestModel> tms) async {
     return new Response.ok(tms);
   }
 }
 
-class OptionalTestController extends RESTController {
+class OptionalTestController extends ResourceController {
   @Operation.post()
   Future<Response> create({@Bind.body() TestModel tm}) async {
     return new Response.ok(tm);
   }
 }
 
-class PropertyTestController extends RESTController {
+class PropertyTestController extends ResourceController {
   @Bind.body()
   TestModel tm;
 
@@ -227,21 +227,21 @@ class PropertyTestController extends RESTController {
   }
 }
 
-class NotSerializableController extends RESTController {
+class NotSerializableController extends ResourceController {
   @Operation.post()
   Future<Response> create(@Bind.body() Uri uri) async {
     return new Response.ok(null);
   }
 }
 
-class ListNotSerializableController extends RESTController {
+class ListNotSerializableController extends ResourceController {
   @Operation.post()
   Future<Response> create(@Bind.body() List<Uri> uri) async {
     return new Response.ok(null);
   }
 }
 
-class CrashController extends RESTController {
+class CrashController extends ResourceController {
   @Operation.post()
   Future<Response> create(@Bind.body() CrashModel tm) async {
     return new Response.ok(null);